project_tier/events/migrations/0002_eventpage_address.py
# -*- coding: utf-8 -*-
from __future__ import unicode_literals

from django.db import migrations, models
import wagtail.core.fields


class Migration(migrations.Migration):

    dependencies = [
        ('events', '0001_initial'),
    ]

    operations = [
        migrations.AddField(
            model_name='eventpage',
            name='address',
            field=wagtail.core.fields.RichTextField(blank=True),
        ),
    ]
