requirements.txt
wagtail==2.0b1
boto==2.48.0
celery==4.1.0
django_compressor==2.2
django-storages==1.6.5
djangorestframework==3.7.7
django-libsass==0.7
django_redis==4.8.0
dj-database-url==0.4.2
wagtailfontawesome==1.1.2
wagtailemoji==1.0.2
wagtailatomicadmin==0.2.5
psycopg2==2.7.3.2
whitenoise==3.3.1

# Tests
django-nose==1.4.5
factory_boy==2.10.0
